requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
